requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure
azure-identity
azure-keyvault-secrets
azure-keyvault
pycryptodome
requests
coinbase==2.1.0
Crypto
python-dotenv